Перенос части логики в сервисы, обработка ошибок

+++ src/main/java/org/nikita111100/onlinetesting/services/PossibleAnswerService.java
@@ -3,16 +3,18 @@ package org.nikita111100.onlinetesting.services;
 import org.nikita111100.onlinetesting.model.persistent.PossibleAnswer;
 import org.nikita111100.onlinetesting.model.persistent.Question;
 import org.nikita111100.onlinetesting.repositories.PossibleAnswerRepo;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.persistence.EntityNotFoundException;
 import java.util.List;
 import java.util.Optional;
 
 @Service
 public class PossibleAnswerService {
 
-//    private static final Logger logger = LoggerFactory.getLogger(PossibleAnswerService.class);
+    //    private static final Logger logger = LoggerFactory.getLogger(PossibleAnswerService.class);
     private final PossibleAnswerRepo pAnswerRepo;
     private final QuestionService questionService;
 
@@ -42,14 +44,44 @@ public class PossibleAnswerService {
     }
 
     @Transactional
-    public PossibleAnswer save(PossibleAnswer possibleAnswer) {
+    public PossibleAnswer save(PossibleAnswer possibleAnswer) throws EntityNotFoundException {
         return pAnswerRepo.save(possibleAnswer);
     }
 
     @Transactional
-    public void deleteById(Long id) {
-        pAnswerRepo.deleteById(id);
+    public void deleteById(Long id) throws EmptyResultDataAccessException {
+        Optional<PossibleAnswer> possibleAnswer = pAnswerRepo.findById(id);
+        if (possibleAnswer.isPresent()) {
+            pAnswerRepo.deleteById(id);
+        }
     }
 
+    @Transactional
+    public void createPossibleAnswerForm(String rolesChecked, PossibleAnswer possibleAnswer, Long questionId) {
+        if (rolesChecked != null) {
+            possibleAnswer.setCorrectAnswer(1);
+        } else {
+            possibleAnswer.setCorrectAnswer(0);
+        }
+        Optional<Question> question = questionService.findById(questionId);
+        if (question.isPresent()) {
+            possibleAnswer.setQuestions(question.get());
+            pAnswerRepo.save(possibleAnswer);
+        }
+    }
+
+    @Transactional
+    public void updatePossibleAnswerForm(String rolesChecked,
+                                         PossibleAnswer possibleAnswer) throws EntityNotFoundException {
+        if (rolesChecked != null) {
+            possibleAnswer.setCorrectAnswer(1);
+        } else {
+            possibleAnswer.setCorrectAnswer(0);
+        }
+        Optional<PossibleAnswer> possibleAnswerFromDb = pAnswerRepo.findById(possibleAnswer.getId());
+        if (possibleAnswerFromDb.isPresent()) {
+            pAnswerRepo.save(possibleAnswer);
+        }
+    }
 
 }
